+++ test/ReadonlyArray.ts
@@ -11,6 +11,7 @@ import * as _ from '../src/ReadonlyArray'
 import * as Show from '../src/Show'
 import * as T from '../src/Task'
 import * as B from '../src/boolean'
+import * as S from '../src/string'
 
 describe('ReadonlyArray', () => {
   describe('pipeables', () => {
@@ -760,8 +761,8 @@ describe('ReadonlyArray', () => {
     assert.deepStrictEqual(_.uniq(Eq.eqNumber)([1, 2, 3, 4, 5]), [1, 2, 3, 4, 5])
     assert.deepStrictEqual(_.uniq(Eq.eqNumber)([1, 1, 2, 2, 3, 3, 4, 4, 5, 5]), [1, 2, 3, 4, 5])
     assert.deepStrictEqual(_.uniq(Eq.eqNumber)([1, 2, 3, 4, 5, 1, 2, 3, 4, 5]), [1, 2, 3, 4, 5])
-    assert.deepStrictEqual(_.uniq(Eq.eqString)(['a', 'b', 'a']), ['a', 'b'])
-    assert.deepStrictEqual(_.uniq(Eq.eqString)(['a', 'b', 'A']), ['a', 'b', 'A'])
+    assert.deepStrictEqual(_.uniq(S.Eq)(['a', 'b', 'a']), ['a', 'b'])
+    assert.deepStrictEqual(_.uniq(S.Eq)(['a', 'b', 'A']), ['a', 'b', 'A'])
 
     assert.strictEqual(_.uniq(Eq.eqNumber)(_.empty), _.empty)
     const as: ReadonlyArray<number> = [1]